+++ CHANGES.md
@@ -1,5 +1,11 @@
 # Release Notes
 
+## 0.1.0 (TBD)
+
+* update FastAPI requirements
+* add `validate` in `MosaicTilerFactory` (https://github.com/developmentseed/titiler/pull/206, author @drnextgis)
+* rename `ressources` package to `resources` (https://github.com/developmentseed/titiler/pull/210, author @drnextgis)
+
 ## 0.1.0a14 (2021-01-05)
 
 * add `rio_tiler.errors.MissingBands` in known errors.